+++ packages/create-vegas/template-preact/vegas.config.ts
@@ -1,0 +1,6 @@
+import { defineConfig } from '@vegasjs/vegas'
+import preact from '@preact/preset-vite'
+
+export default defineConfig({
+  plugins: [preact()],
+})