+++ packages/create-vegas/template-vue/src/web/App.vue
@@ -1,7 +1,7 @@
-<script setup lang="ts"></script>
+<script setup lang="ts">
+import HelloWorld from "./components/HelloWorld.vue";
+</script>
 
 <template>
-  <h1>Hello, world!</h1>
+  <HelloWorld />
 </template>
-
-<style scoped></style>