+++ packages/create-vegas/template-vue/src/web/main.tsx
@@ -1,5 +1,6 @@
-import { createApp } from "vue";
+import { createApp } from "vue"
+import './style.css'
 
-import App from "./App.vue";
+import App from "./App.vue"
 
-createApp(App).mount("#root");
+createApp(App).mount("#root")