+++ packages/vegas/src/node/cli/serve/handlers/urlFetchApp.ts
@@ -13,4 +13,28 @@ export class UrlFetchAppHandler {
       responseCode: response.status,
     };
   }
+  async fetchAll(
+    _ctx: ServeContext,
+    payload: { fetchRequests: { url: string; init?: RequestInit }[] },
+  ) {
+    const responses = await Promise.all(
+      payload.fetchRequests.map((fetchRequest) => {
+        return fetch(fetchRequest.url, fetchRequest.init);
+      }),
+    );
+
+    return await Promise.all(
+      responses.map(async (response) => {
+        const headers: Record<string, string> = {};
+        response.headers.forEach((value, key) => {
+          headers[key] = value;
+        });
+        return {
+          headers,
+          content: Array.from(await response.bytes()),
+          responseCode: response.status,
+        };
+      }),
+    );
+  }
 }